Refactor global configuration reset handling
