+++ assertj-core/src/test/java/org/assertj/core/api/EntryPointAssertions_setAllowComparingPrivateFields_Test.java
@@ -21,7 +21,6 @@ import java.util.function.Consumer;
 import java.util.stream.Stream;
 
 import org.assertj.core.util.introspection.FieldSupport;
-import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.MethodSource;
@@ -30,12 +29,6 @@ import org.junit.jupiter.params.provider.MethodSource;
 class EntryPointAssertions_setAllowComparingPrivateFields_Test extends EntryPointAssertionsBaseTest {
 
   private static final boolean DEFAULT_ALLOW_COMPARING_PRIVATE_FIELDS = FieldSupport.comparison().isAllowedToUsePrivateFields();
-
-  @AfterEach
-  void afterEachTest() {
-    // reset to the default value to avoid side effects on the other tests
-    FieldSupport.comparison().setAllowUsingPrivateFields(DEFAULT_ALLOW_COMPARING_PRIVATE_FIELDS);
-  }
 
   @ParameterizedTest
   @MethodSource("setAllowComparingPrivateFieldsFunctions")
