+++ assertj-core/src/test/java/org/assertj/core/api/EntryPointAssertions_setMaxLengthForSingleLineDescription_Test.java
@@ -21,19 +21,12 @@ import java.util.function.Consumer;
 import java.util.stream.Stream;
 
 import org.assertj.core.presentation.StandardRepresentation;
-import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.MethodSource;
 
 class EntryPointAssertions_setMaxLengthForSingleLineDescription_Test extends EntryPointAssertionsBaseTest {
 
   private static final int DEFAULT_MAX_LENGTH_FOR_SINGLE_LINE = StandardRepresentation.getMaxLengthForSingleLineDescription();
-
-  @AfterEach
-  void afterEachTest() {
-    // reset to the default value to avoid side effects on the other tests
-    StandardRepresentation.setMaxLengthForSingleLineDescription(DEFAULT_MAX_LENGTH_FOR_SINGLE_LINE);
-  }
 
   @ParameterizedTest
   @MethodSource("setMaxLengthForSingleLineDescriptionFunctions")
