+++ assertj-core/src/test/java/org/assertj/core/api/EntryPointAssertions_useDefaultRepresentation_Test.java
@@ -24,23 +24,12 @@ import java.util.stream.Stream;
 
 import org.apache.commons.lang3.tuple.Pair;
 import org.assertj.core.presentation.Representation;
-import org.assertj.core.testkit.MutatesGlobalConfiguration;
-import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.MethodSource;
 
 @DisplayName("EntryPoint assertions useDefaultRepresentation method")
-@MutatesGlobalConfiguration
 class EntryPointAssertions_useDefaultRepresentation_Test extends EntryPointAssertionsBaseTest {
-
-  private static final Representation DEFAULT_CUSTOM_REPRESENTATION = AbstractAssert.customRepresentation;
-
-  @AfterEach
-  void afterEachTest() {
-    // reset to the default value to avoid side effects on the other tests
-    AbstractAssert.customRepresentation = DEFAULT_CUSTOM_REPRESENTATION;
-  }
 
   @ParameterizedTest
   @MethodSource("useRepresentationFunctions")
