+++ assertj-core/src/test/java/org/assertj/core/api/WithAssertions_delegation_Test.java
@@ -64,6 +64,7 @@ import java.util.function.Predicate;
 import java.util.stream.Stream;
 
 import org.assertj.core.data.MapEntry;
+import org.assertj.core.internal.Failures;
 import org.assertj.core.util.Lists;
 import org.junit.jupiter.api.Test;
 
@@ -761,7 +762,12 @@ class WithAssertions_delegation_Test implements WithAssertions {
    */
   @Test
   void withAssertions_setRemoveAssertJRelatedElementsFromStackTrace_Test() {
-    setRemoveAssertJRelatedElementsFromStackTrace(true);
+    boolean removeAssertJRelatedElementsFromStackTrace = Failures.instance().isRemoveAssertJRelatedElementsFromStackTrace();
+    try {
+      setRemoveAssertJRelatedElementsFromStackTrace(true);
+    } finally {
+      setRemoveAssertJRelatedElementsFromStackTrace(removeAssertJRelatedElementsFromStackTrace);
+    }
   }
 
   /**
