+++ assertj-tests/assertj-integration-tests/assertj-core-tests/src/test/java/module-info.java
@@ -30,4 +30,6 @@ open module org.assertj.tests.core {
   requires org.opentest4j;
   requires org.junit.jupiter.api;
   requires org.apache.commons.collections4;
+
+  provides org.junit.jupiter.api.extension.Extension with org.assertj.tests.core.testkit.ResetGlobalConfigurationExtension;
 }
