+++ assertj-tests/assertj-integration-tests/assertj-core-tests/src/test/resources/META-INF/services/org.junit.jupiter.api.extension.Extension
@@ -1,0 +1,1 @@
+org.assertj.tests.core.testkit.ResetGlobalConfigurationExtension
