+++ assertj-tests/assertj-integration-tests/assertj-core-tests/src/test/resources/junit-platform.properties
@@ -1,1 +1,2 @@
 junit.jupiter.displayname.generator.default=org.assertj.tests.core.testkit.junit.jupiter.DefaultDisplayNameGenerator
+junit.jupiter.extensions.autodetection.enabled=true
